pyproject.toml
[tool.poetry]
name = "morning-bot"
version = "0.1.0"
description = ""
authors = ["alina1412 <alina1412>"]
readme = "README.md"
packages = [{include = "morning_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
aiofiles = "0.8.0"
anyio = "3.5.0"
asyncio = "3.4.3"
atomicwrites = "1.4.0"
attrs = "21.4.0"
certifi = "2021.10.8"
charset-normalizer = "2.0.12"
colorama = "0.4.5"
h11 = "0.12.0"
httpcore = "0.14.7"
httpx = "0.22.0"
idna = "3.3"
iniconfig = "1.1.1"
packaging = "21.3"
pluggy = "1.0.0"
py = "1.11.0"
pyparsing = "3.0.9"
requests = "2.28.1"
rfc3986 = "1.5.0"
sniffio = "1.2.0"
tomli = "2.0.1"
urllib3 = "1.26.9"
python-dotenv = "^0.21.0"
uvicorn = "^0.19.0"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
pylint = "^2.15.5"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

asyncio_mode = "auto"